% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{Binomial}
\alias{Binomial}
\title{Binomial Estimate}
\usage{
Binomial(data, cL, trials)
}
\arguments{
\item{data}{Data set/sample from distribution}

\item{trials}{number of trials}

\item{CL}{confidence interval}
}
\value{
estimate of probability, confidence interval
}
\description{
Estimates of parameter p of a binomial distribution & its confidence interval
n always 1.
}
\examples{
Binomial(rbinom(1,10,0.49),.95,10)
Binomial(rbinom(1,20,0.7),.95,20)
}
