% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRD.R
\name{crd}
\alias{crd}
\title{This function performs Complete Randomize Design on a given data set.}
\arguments{
\item{a}{Parameter a is the number of treatment levels for the given experiment.}

\item{n}{Parameter n is the total number of observation/response for the given experiment.}

\item{response}{Response is a vector of length n to perform the CRD on.}

\item{alp}{Parameter alp is the level of significance. This is required for confidence intervals and drawing conclusions.}
}
\value{
This function returns:
\enumerate{
\item The resulting ANOVA table for CRD,
\item Decision i.e. whether to reject or accept Null hypothesis,
\item Result of the tukey's test,
\item Confidence interval for every treatment mean,
\item Limits of the Bonferroni's test.
}
}
\description{
This function performs Complete Randomize Design on a given data set.
}
\examples{
crd(a = # number of treatment levels,n = #total number of response,  response= #Dataset as vector of length n, alp = # significance level )
}
