\name{Design-internal}
\title{Internal Design functions}
\alias{addOffset4ModelFrame}
\alias{as.character.Surv}
\alias{axisf}
\alias{bj.fit2}
\alias{coxphFit}
\alias{des.args}
\alias{[.Design}
\alias{DesignAssign}
\alias{Design.levels}
\alias{formula.Design}
\alias{getOldDesign}
\alias{gparms}
\alias{is.na.Surv}
\alias{is.Surv}
\alias{lm.pfit}
\alias{lrm.fit.strat}
\alias{Math.Surv}
\alias{oldDesignFit2R}
\alias{ols.influence}
\alias{Ops.Surv}
\alias{print.summary.survreg2}
\alias{set.atr}
\alias{summary.glmD}
\alias{Summary.Surv}
\alias{Surv}
\alias{[.Surv}
\alias{survfit.cph.null}
\alias{survreg.auxinfo}
\alias{survreg.fit2}
\alias{val.probg}
\alias{value.chk}
\alias{.R.}
\alias{.SV4.}
\alias{.newSurvival.}
\alias{strgraphwrap}
\description{Internal Design functions.}
\details{These are not to be called by the user or are undocumented.}
\keyword{internal}
