% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_function.R
\name{design_function}
\alias{design_function}
\title{A design Function}
\usage{
design_function(love = TRUE)
}
\arguments{
\item{love}{Do you love design? Defaults to TRUE.}
}
\description{
This function allows you to express your love for design.
}
\examples{
design_function()
}
\keyword{design}
