% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtlcor_sim.R
\name{dtl_app_sim}
\alias{dtl_app_sim}
\title{simulation function}
\usage{
dtl_app_sim(
  nsim,
  alpha_t,
  D,
  N,
  n,
  mPFS,
  q,
  gamma,
  delta,
  drop_rate,
  follow_time,
  enroll,
  t
)
}
\description{
simulation function
}
