% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtlcor_fun.R
\name{dtl_cor_the_PH_upper_bound}
\alias{dtl_cor_the_PH_upper_bound}
\title{get theoretical upper bound of correlation coefficient}
\usage{
dtl_cor_the_PH_upper_bound(tau_k, pi_ar = 0.5, q, gamma)
}
\description{
get theoretical upper bound of correlation coefficient
}
