% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exponential.R
\name{exponential}
\alias{exponential}
\title{Exponential Estimate}
\usage{
exponential(data, cL)
}
\arguments{
\item{data}{A Data set/Sample}

\item{cL}{Confidence interval}
}
\value{
estimated parameter lamda,confidence interval
}
\description{
Estimation of the parameter & confidence interval from an exponential distribution
}
\examples{
exponential(rexp(1000,5),.95)
exponential(rexp(10000,34),.98)
}
