% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSD.R
\name{LSD}
\alias{LSD}
\title{Performs Latin Square design on a given data set.}
\usage{
LSD(x)
}
\arguments{
\item{x}{The provided data set to perform Latin square design on.
It is to be remembered that in Latin square design, the number of columns, rows and treatments should be equal, otherwise it'll theoretically be inappropriate.}
}
\value{
Returns the Analysis of variance table with the F-values with respect to the treatment, row factor and column factor.
}
\description{
Performs Latin Square design on a given data set.
}
\examples{
LSD(x = #Dataset as tibble / dataframe)
}
