\name{print.lrm}
\alias{print.lrm}
\title{
Print lrm
}
\description{
Formatted printing of an object of class \code{lrm}
}
\usage{
\method{print}{lrm}(x, digits=4, strata.coefs=FALSE, \dots)
}
\arguments{
\item{x}{fit object}
\item{digits}{number of significant digits to use}
\item{strata.coefs}{set to \code{TRUE} to print the (experimental)
  strata coefficients}
\item{\dots}{ignored}
}
\keyword{print}
