% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCBD.R
\name{rcbd}
\alias{rcbd}
\title{This function performs Randomized complete block design on a given data set.}
\arguments{
\item{data}{Data is the data frame to perform RCBD on. Input should be a data frame that contains 3 columns named - raw_data, treat & block}

\item{alp}{Alpha is the level of significance. This is required to compute the confidence level and drawing conclusions.}
}
\value{
This function returns the ANOVA table, Decision i.e. to accept or reject the null hypothesis, and two plots : comparative box-plot and  Normal Q-Q plot.
}
\description{
This function performs Randomized complete block design on a given data set.
}
\examples{
rcbd(data = #Dataset as tibble / dataframe, alp = # significance level)
}
