% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSD.R
\name{sumsq}
\alias{sumsq}
\title{Calculates the sum of the squares of the elements of a provided vector.}
\usage{
sumsq(x)
}
\arguments{
\item{x}{A vector}
}
\value{
Outputs the sum of the squares of the elements of the supplied vector.
}
\description{
Calculates the sum of the squares of the elements of a provided vector.
}
\examples{
sumsq(c(1:5))
}
