% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_of_hypothesis.R
\name{test_of_hypothesis}
\alias{test_of_hypothesis}
\title{Test of hypothesis which determines the p-value,confidence interval and test
statistic given the arguments}
\usage{
test_of_hypothesis(
  parameter,
  independent = NULL,
  x,
  y = NULL,
  mu_0 = NULL,
  sigma = NULL,
  n1 = NULL,
  n2 = NULL,
  B = NULL,
  alpha,
  H1
)
}
\arguments{
\item{parameter}{argument to determine whether the parameter is known}

\item{independent}{For matched pair data set, this parameter is false, and otherwise true.}

\item{x}{A random variable from a normal distribution}

\item{y}{Another random variable from a distribution}

\item{mu_0}{mean of a specific distribution}

\item{sigma}{standard of a specific distribution}

\item{n1}{number of rows matrix}

\item{n2}{number of columns matrix}

\item{B}{number of replication}

\item{alpha}{level of significance}

\item{H1}{alternate hypothesis}
}
\value{
p-value,confidence-interval,test-statistic
}
\description{
Test of hypothesis which determines the p-value,confidence interval and test
statistic given the arguments
}
\examples{
test_of_hypothesis(parameter = "known", independent = FALSE, x = x,
y = y, mu_0 = mu_0, alpha = alpha, H1 = H1 )
}
