% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtlcor_sim.R
\name{dtl_app_sim_single}
\alias{dtl_app_sim_single}
\title{simulate a single trial}
\usage{
dtl_app_sim_single(
  D,
  N,
  n,
  mPFS,
  q,
  gamma,
  delta,
  drop_rate,
  follow_time,
  enroll,
  t
)
}
\description{
simulate a single trial
}
